changed all network-generators to work on int-edgeweights

+++ src/dna/graph/generators/network/m1/M1Batch.java
@@ -51,8 +51,9 @@ public class M1Batch extends NetworkBatch {
 	}
 
 	@Override
-	public Batch craftBatch(Graph g, DateTime timestamp, ArrayList<TCPEvent> events,
-			HashMap<String, Long> weightChangesMap) {
+	public Batch craftBatch(Graph g, DateTime timestamp,
+			ArrayList<TCPEvent> events,
+			HashMap<String, Integer> weightChangesMap) {
 		ArrayList<String> addedNodes = new ArrayList<String>();
 		ArrayList<Node> addedNodesNodes = new ArrayList<Node>();
 		ArrayList<NetworkEdge> addedEdges = new ArrayList<NetworkEdge>();
